LayoutTests/platform/blackberry/fast/dom/Window/dom-access-from-closure-window-expected.txt

CONSOLE MESSAGE: line 4: TypeError: Result of expression 'opener.done' [undefined] is not a function.
FAIL: Timed out waiting for notifyDone to be called
